Move Widget trait into widget.rs, add box_constraints.rs
