+++ druid/src/core.rs
@@ -14,139 +14,13 @@
 
 //! The fundamental druid types.
 
-use log;
 use std::collections::VecDeque;
 use std::ops::{Deref, DerefMut};
 use std::time::Instant;
 
 use crate::kurbo::{Rect, Size};
 use crate::piet::Piet;
-use crate::{
-    BaseState, Command, Cursor, Env, Event, Text, TimerToken, WinCtx, WindowHandle, WindowId,
-};
-
-/// The trait implemented by all widgets.
-///
-/// All appearance and behavior for a widget is encapsulated in an
-/// object that implements this trait.
-///
-/// The trait is parametrized by a type (`T`) for associated data.
-/// All trait methods are provided with access to this data, and
-/// in the case of `event` the reference is mutable, so that events
-/// can directly update the data.
-///
-/// Whenever the application data changes, the framework traverses
-/// the widget hierarchy with an [`update`] method. The framework
-/// needs to know whether the data has actually changed or not, which
-/// is why `T` has a [`Data`] bound.
-///
-/// All the trait methods are provided with a corresponding context.
-/// The widget can request things and cause actions by calling methods
-/// on that context.
-///
-/// In addition, all trait methods are provided with an environment
-/// ([`Env`](struct.Env.html)).
-///
-/// Container widgets will generally not call `Widget` methods directly
-/// on their child widgets, but rather will own their widget wrapped in
-/// a [`WidgetPod`], and call the corresponding method on that. The
-/// `WidgetPod` contains state and logic for these traversals. On the
-/// other hand, particularly light-weight containers might contain their
-/// child `Widget` directly (when no layout or event flow logic is
-/// needed), and in those cases will call these methods.
-///
-/// As a general pattern, container widgets will call the corresponding
-/// `WidgetPod` method on all their children. The `WidgetPod` applies
-/// logic to determine whether to recurse, as needed.
-///
-/// [`event`]: #tymethod.event
-/// [`update`]: #tymethod.update
-/// [`Data`]: trait.Data.html
-/// [`WidgetPod`]: struct.WidgetPod.html
-pub trait Widget<T> {
-    /// Handle an event.
-    ///
-    /// A number of different events (in the [`Event`] enum) are handled in this
-    /// method call. A widget can handle these events in a number of ways:
-    /// requesting things from the [`EventCtx`], mutating the data, or submitting
-    /// a [`Command`].
-    ///
-    /// [`Event`]: struct.Event.html
-    /// [`EventCtx`]: struct.EventCtx.html
-    /// [`Command`]: struct.Command.html
-    fn event(&mut self, ctx: &mut EventCtx, event: &Event, data: &mut T, env: &Env);
-
-    /// Handle a change of data.
-    ///
-    /// This method is called whenever the data changes. When the appearance of
-    /// the widget depends on data, call [`invalidate`] so that it's scheduled
-    /// for repaint.
-    ///
-    /// The previous value of the data is provided in case the widget wants to
-    /// compute a fine-grained delta. Before any paint operation, this method
-    /// will be called with `None` for `old_data`. Thus, this method can also be
-    /// used to build resources that will be retained for painting.
-    ///
-    /// [`invalidate`]: struct.UpdateCtx.html#method.invalidate
-
-    // Consider a no-op default impl. One reason against is that containers might
-    // inadvertently forget to propagate.
-    fn update(&mut self, ctx: &mut UpdateCtx, old_data: Option<&T>, data: &T, env: &Env);
-
-    /// Compute layout.
-    ///
-    /// A leaf widget should determine its size (subject to the provided
-    /// constraints) and return it.
-    ///
-    /// A container widget will recursively call [`WidgetPod::layout`] on its
-    /// child widgets, providing each of them an appropriate box constraint,
-    /// compute layout, then call [`set_layout_rect`] on each of its children.
-    /// Finally, it should return the size of the container. The container
-    /// can recurse in any order, which can be helpful to, for example, compute
-    /// the size of non-flex widgets first, to determine the amount of space
-    /// available for the flex widgets.
-    ///
-    /// For efficiency, a container should only invoke layout of a child widget
-    /// once, though there is nothing enforcing this.
-    ///
-    /// The layout strategy is strongly inspired by Flutter.
-    ///
-    /// [`WidgetPod::layout`]: struct.WidgetPod.html#method.layout
-    /// [`set_layout_rect`]: struct.LayoutCtx.html#method.set_layout_rect
-    fn layout(&mut self, ctx: &mut LayoutCtx, bc: &BoxConstraints, data: &T, env: &Env) -> Size;
-
-    /// Paint the widget appearance.
-    ///
-    /// The widget calls methods on the `render_ctx` field of the
-    /// `paint_ctx` in order to paint its appearance. `paint_ctx` auto
-    /// derefs to `render_ctx` for convenience.
-    ///
-    /// Container widgets can paint a background before recursing to their
-    /// children, or annotations (for example, scrollbars) by painting
-    /// afterwards. In addition, they can apply masks and transforms on
-    /// the render context, which is especially useful for scrolling.
-    fn paint(&mut self, paint_ctx: &mut PaintCtx, base_state: &BaseState, data: &T, env: &Env);
-}
-
-// TODO: explore getting rid of this (ie be consistent about using
-// `dyn Widget` only).
-impl<T> Widget<T> for Box<dyn Widget<T>> {
-    fn event(&mut self, ctx: &mut EventCtx, event: &Event, data: &mut T, env: &Env) {
-        self.deref_mut().event(ctx, event, data, env)
-    }
-
-    fn update(&mut self, ctx: &mut UpdateCtx, old_data: Option<&T>, data: &T, env: &Env) {
-        self.deref_mut().update(ctx, old_data, data, env);
-    }
-
-    fn layout(&mut self, ctx: &mut LayoutCtx, bc: &BoxConstraints, data: &T, env: &Env) -> Size {
-        self.deref_mut().layout(ctx, bc, data, env)
-    }
-
-    fn paint(&mut self, paint_ctx: &mut PaintCtx, base_state: &BaseState, data: &T, env: &Env) {
-        self.deref_mut().paint(paint_ctx, base_state, data, env);
-    }
-}
+use crate::{BaseState, Command, Cursor, Text, TimerToken, WinCtx, WindowHandle, WindowId};
 
 /// A context passed to paint methods of widgets.
 ///
@@ -276,109 +150,6 @@ pub struct UpdateCtx<'a, 'b: 'a> {
     // now keep it super-simple.
     pub(crate) needs_inval: bool,
     pub(crate) window_id: WindowId,
-}
-
-/// Constraints for layout.
-///
-/// The layout strategy for druid is strongly inspired by Flutter,
-/// and this struct is similar to the [Flutter BoxConstraints] class.
-///
-/// At the moment, it represents simply a minimum and maximum size.
-/// A widget's [`layout`] method should choose an appropriate size that
-/// meets these constraints.
-///
-/// Further, a container widget should compute appropriate constraints
-/// for each of its child widgets, and pass those down when recursing.
-///
-/// [`layout`]: trait.Widget.html#tymethod.layout
-/// [Flutter BoxConstraints]: https://api.flutter.dev/flutter/rendering/BoxConstraints-class.html
-#[derive(Clone, Copy, Debug)]
-pub struct BoxConstraints {
-    min: Size,
-    max: Size,
-}
-
-impl BoxConstraints {
-    /// Create a new box constraints object.
-    ///
-    /// Create constraints based on minimum and maximum size.
-    pub fn new(min: Size, max: Size) -> BoxConstraints {
-        BoxConstraints { min, max }
-    }
-
-    /// Create a "tight" box constraints object.
-    ///
-    /// A "tight" constraint can only be satisfied by a single size.
-    pub fn tight(size: Size) -> BoxConstraints {
-        BoxConstraints {
-            min: size,
-            max: size,
-        }
-    }
-
-    /// Create a "loose" version of the constraints.
-    ///
-    /// Make a version with zero minimum size, but the same maximum size.
-    pub fn loosen(&self) -> BoxConstraints {
-        BoxConstraints {
-            min: Size::ZERO,
-            max: self.max,
-        }
-    }
-
-    /// Clamp a given size so that fits within the constraints.
-    pub fn constrain(&self, size: impl Into<Size>) -> Size {
-        size.into().clamp(self.min, self.max)
-    }
-
-    /// Returns the max size of these constraints.
-    pub fn max(&self) -> Size {
-        self.max
-    }
-
-    /// Returns the min size of these constraints.
-    pub fn min(&self) -> Size {
-        self.min
-    }
-
-    /// Whether there is an upper bound on the width.
-    pub fn is_width_bounded(&self) -> bool {
-        self.max.width.is_finite()
-    }
-
-    /// Whether there is an upper bound on the height.
-    pub fn is_height_bounded(&self) -> bool {
-        self.max.height.is_finite()
-    }
-
-    /// Check to see if these constraints are legit.
-    ///
-    /// Logs a warning if BoxConstraints are invalid.
-    pub fn debug_check(&self, name: &str) {
-        if !(0.0 <= self.min.width
-            && self.min.width <= self.max.width
-            && 0.0 <= self.min.height
-            && self.min.height <= self.max.height)
-        {
-            log::warn!("Bad BoxConstraints passed to {}:", name);
-            log::warn!("{:?}", self);
-        }
-    }
-
-    /// Shrink min and max constraints by size
-    pub fn shrink(&self, diff: impl Into<Size>) -> BoxConstraints {
-        let diff = diff.into();
-        let min = Size::new(
-            (self.min().width - diff.width).max(0.),
-            (self.min().height - diff.height).max(0.),
-        );
-        let max = Size::new(
-            (self.max().width - diff.width).max(0.),
-            (self.max().height - diff.height).max(0.),
-        );
-
-        BoxConstraints::new(min, max)
-    }
 }
 
 impl<'a, 'b> EventCtx<'a, 'b> {
