+++ druid/src/widget/mod.rs
@@ -15,68 +15,177 @@
 //! Common widgets.
 
 mod align;
-pub use crate::widget::align::Align;
-
 mod button;
-pub use crate::widget::button::Button;
-
-mod label;
-pub use crate::widget::label::{Label, LabelText};
-
-mod either;
-pub use crate::widget::either::Either;
-
-mod flex;
-pub use crate::widget::flex::{Column, Flex, Row};
-
-mod padding;
-pub use crate::widget::padding::Padding;
-
-mod scroll;
-pub use crate::widget::scroll::Scroll;
-
-mod progress_bar;
-pub use crate::widget::progress_bar::ProgressBar;
-
-mod slider;
-pub use crate::widget::slider::Slider;
-
-mod textbox;
-pub use crate::widget::textbox::TextBox;
-
-mod sized_box;
-pub use crate::widget::sized_box::SizedBox;
-
 mod checkbox;
-pub use crate::widget::checkbox::Checkbox;
-
-mod radio;
-pub use crate::widget::radio::{Radio, RadioGroup};
-
 mod container;
-pub use crate::widget::container::Container;
-
-mod split;
-pub use crate::widget::split::Split;
-
-mod switch;
-pub use crate::widget::switch::Switch;
-
+mod either;
 mod env_scope;
-pub use crate::widget::env_scope::EnvScope;
-
-mod widget_ext;
-pub use widget_ext::WidgetExt;
-
+mod flex;
+mod label;
 mod list;
-pub use crate::widget::list::{List, ListIter};
-
+mod padding;
 mod parse;
-pub use crate::widget::parse::Parse;
-
+mod progress_bar;
+mod radio;
+mod scroll;
+mod sized_box;
+mod slider;
+mod split;
 #[cfg(feature = "svg")]
 #[cfg_attr(docsrs, doc(cfg(feature = "svg")))]
 mod svg;
+mod switch;
+mod textbox;
+mod widget_ext;
+
+pub use align::Align;
+pub use button::Button;
+pub use checkbox::Checkbox;
+pub use container::Container;
+pub use either::Either;
+pub use env_scope::EnvScope;
+pub use flex::{Column, Flex, Row};
+pub use label::{Label, LabelText};
+pub use list::{List, ListIter};
+pub use padding::Padding;
+pub use parse::Parse;
+pub use progress_bar::ProgressBar;
+pub use radio::{Radio, RadioGroup};
+pub use scroll::Scroll;
+pub use sized_box::SizedBox;
+pub use slider::Slider;
+pub use split::Split;
 #[cfg(feature = "svg")]
 #[cfg_attr(docsrs, doc(cfg(feature = "svg")))]
-pub use crate::widget::svg::{Svg, SvgData};
+pub use svg::{Svg, SvgData};
+pub use switch::Switch;
+pub use textbox::TextBox;
+pub use widget_ext::WidgetExt;
+
+use std::ops::DerefMut;
+
+use crate::kurbo::Size;
+use crate::{BaseState, BoxConstraints, Env, Event, EventCtx, LayoutCtx, PaintCtx, UpdateCtx};
+
+/// The trait implemented by all widgets.
+///
+/// All appearance and behavior for a widget is encapsulated in an
+/// object that implements this trait.
+///
+/// The trait is parametrized by a type (`T`) for associated data.
+/// All trait methods are provided with access to this data, and
+/// in the case of `event` the reference is mutable, so that events
+/// can directly update the data.
+///
+/// Whenever the application data changes, the framework traverses
+/// the widget hierarchy with an [`update`] method. The framework
+/// needs to know whether the data has actually changed or not, which
+/// is why `T` has a [`Data`] bound.
+///
+/// All the trait methods are provided with a corresponding context.
+/// The widget can request things and cause actions by calling methods
+/// on that context.
+///
+/// In addition, all trait methods are provided with an environment
+/// ([`Env`](struct.Env.html)).
+///
+/// Container widgets will generally not call `Widget` methods directly
+/// on their child widgets, but rather will own their widget wrapped in
+/// a [`WidgetPod`], and call the corresponding method on that. The
+/// `WidgetPod` contains state and logic for these traversals. On the
+/// other hand, particularly light-weight containers might contain their
+/// child `Widget` directly (when no layout or event flow logic is
+/// needed), and in those cases will call these methods.
+///
+/// As a general pattern, container widgets will call the corresponding
+/// `WidgetPod` method on all their children. The `WidgetPod` applies
+/// logic to determine whether to recurse, as needed.
+///
+/// [`event`]: #tymethod.event
+/// [`update`]: #tymethod.update
+/// [`Data`]: trait.Data.html
+/// [`WidgetPod`]: struct.WidgetPod.html
+pub trait Widget<T> {
+    /// Handle an event.
+    ///
+    /// A number of different events (in the [`Event`] enum) are handled in this
+    /// method call. A widget can handle these events in a number of ways:
+    /// requesting things from the [`EventCtx`], mutating the data, or submitting
+    /// a [`Command`].
+    ///
+    /// [`Event`]: struct.Event.html
+    /// [`EventCtx`]: struct.EventCtx.html
+    /// [`Command`]: struct.Command.html
+    fn event(&mut self, ctx: &mut EventCtx, event: &Event, data: &mut T, env: &Env);
+
+    /// Handle a change of data.
+    ///
+    /// This method is called whenever the data changes. When the appearance of
+    /// the widget depends on data, call [`invalidate`] so that it's scheduled
+    /// for repaint.
+    ///
+    /// The previous value of the data is provided in case the widget wants to
+    /// compute a fine-grained delta. Before any paint operation, this method
+    /// will be called with `None` for `old_data`. Thus, this method can also be
+    /// used to build resources that will be retained for painting.
+    ///
+    /// [`invalidate`]: struct.UpdateCtx.html#method.invalidate
+
+    // Consider a no-op default impl. One reason against is that containers might
+    // inadvertently forget to propagate.
+    fn update(&mut self, ctx: &mut UpdateCtx, old_data: Option<&T>, data: &T, env: &Env);
+
+    /// Compute layout.
+    ///
+    /// A leaf widget should determine its size (subject to the provided
+    /// constraints) and return it.
+    ///
+    /// A container widget will recursively call [`WidgetPod::layout`] on its
+    /// child widgets, providing each of them an appropriate box constraint,
+    /// compute layout, then call [`set_layout_rect`] on each of its children.
+    /// Finally, it should return the size of the container. The container
+    /// can recurse in any order, which can be helpful to, for example, compute
+    /// the size of non-flex widgets first, to determine the amount of space
+    /// available for the flex widgets.
+    ///
+    /// For efficiency, a container should only invoke layout of a child widget
+    /// once, though there is nothing enforcing this.
+    ///
+    /// The layout strategy is strongly inspired by Flutter.
+    ///
+    /// [`WidgetPod::layout`]: struct.WidgetPod.html#method.layout
+    /// [`set_layout_rect`]: struct.LayoutCtx.html#method.set_layout_rect
+    fn layout(&mut self, ctx: &mut LayoutCtx, bc: &BoxConstraints, data: &T, env: &Env) -> Size;
+
+    /// Paint the widget appearance.
+    ///
+    /// The widget calls methods on the `render_ctx` field of the
+    /// `paint_ctx` in order to paint its appearance. `paint_ctx` auto
+    /// derefs to `render_ctx` for convenience.
+    ///
+    /// Container widgets can paint a background before recursing to their
+    /// children, or annotations (for example, scrollbars) by painting
+    /// afterwards. In addition, they can apply masks and transforms on
+    /// the render context, which is especially useful for scrolling.
+    fn paint(&mut self, paint_ctx: &mut PaintCtx, base_state: &BaseState, data: &T, env: &Env);
+}
+
+// TODO: explore getting rid of this (ie be consistent about using
+// `dyn Widget` only).
+impl<T> Widget<T> for Box<dyn Widget<T>> {
+    fn event(&mut self, ctx: &mut EventCtx, event: &Event, data: &mut T, env: &Env) {
+        self.deref_mut().event(ctx, event, data, env)
+    }
+
+    fn update(&mut self, ctx: &mut UpdateCtx, old_data: Option<&T>, data: &T, env: &Env) {
+        self.deref_mut().update(ctx, old_data, data, env);
+    }
+
+    fn layout(&mut self, ctx: &mut LayoutCtx, bc: &BoxConstraints, data: &T, env: &Env) -> Size {
+        self.deref_mut().layout(ctx, bc, data, env)
+    }
+
+    fn paint(&mut self, paint_ctx: &mut PaintCtx, base_state: &BaseState, data: &T, env: &Env) {
+        self.deref_mut().paint(paint_ctx, base_state, data, env);
+    }
+}
